Dynamo-Server/requirements.txt
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.1
decorator==4.0.11
delegator.py==0.1.1
Django==2.2.6
django-discover-runner==1.0
django-jet==1.0.7
django-picklefield==0.3.2
django-role-permissions==2.2.0
django-sendfile==0.3.11
djangorestframework==3.8.2
djangorestframework-jwt==1.10.0
idna==2.8
packaging==16.8
pexpect==4.7.0
pkg-resources==0.0.0
psycopg2==2.8.4
ptyprocess==0.6.0
PyJWT==1.7.1
pyparsing==2.4.2
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
